requirements.txt
ccxt
pandas
ta
openpyxl
python-dotenv
